refactor: Move Steps to components
